This week Top Read Article in academia....!!!!
#gps #accuracyfactors #improveaccuracy #globalpositioningsystem #globalsystem #ionosphere #troposphereregions #stationsetworks #algorithm #mapmatchingalgorithm #malaysia

Paper Title
GPS SYSTEMS LITERATURE: INACCURACY FACTORS AND EFFECTIVE SOLUTIONS

Author Details
Li Nyen Thin, Lau Ying Ting, Nor Adila Husna and Mohd Heikal Husin, Universiti Sains Malaysia, Malaysia

ABSTRACT
Today, Global Positioning System (GPS) is widely used in almost every aspect of our daily life. Commonly, users utilize the technology to track the position of a vehicle or an object of interest. They also use it to safely navigate to the destination of their choice. As a result, there are countless number of GPS based tracking application that has been developed. But, a main recurring issue that exists among these applications are the inaccuracy of the tracking faced by users and this issue has become a rising concern. Most existing research have examined the effects that the inaccuracy of GPS have on users while others identified suitable methods to improve the accuracy of GPS based on one or two factors. The objective of this survey paper is to identify the common factors that affects the accuracy of GPS and identify an effective method which could mitigate or overcome most of those factors. As part of our research, we conducted a thorough examination of the existing factors for GPS inaccuracies. According to an initial survey that we have collected, most of the respondents has faced some form of GPS inaccuracy. Among the common issues faced are inaccurate object tracking and disconnection of GPS signal while using an application. As such, most of the respondents agree that it is necessary to improve the accuracy of GPS. This leads to another objective of this paper, which is to examine and evaluate existing methods as well as to identify the most effective method that could improve the accuracy of GPS.

KEYWORDS
GPS, accuracy factors, improve accuracy, global positioning system 

Pdf URL: https://aircconline.com/ijcnc/V8N2/8216cnc11.pdf

https://www.academia.edu/24447613/GPS_SYSTEMS_LITERATURE_INACCURACY_FACTORS_AND_EFFECTIVE_SOLUTIONS